CODE-16 - Moving from JPA to JDBCTemplate and JdbcClient (Spring Boot 3.3.x)
- Update Spring boot 3.3.1
- Added RestClient
- Using bundle for Certificates
- Fixed some tests
- Updated README.md

+++ novel-ideas-rest/src/main/java/com/klemmy/novelideas/jpa/repository/BookRepository.java
@@ -17,11 +17,11 @@ import java.util.Optional;
 import static org.springframework.data.jpa.domain.Specification.where;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer>, JpaSpecificationExecutor<Book> {
+public interface BookRepository extends JpaRepository<Book, Long>, JpaSpecificationExecutor<Book> {
 
   void delete(Book book);
 
-  Optional<List<CharacterProfile>> findBookById(Integer id);
+  Optional<List<CharacterProfile>> findBookById(Long id);
 
   default Page<Book> findAllByFilters(final String queryTitle,
                                       final LocalDateTime startDate,
